refactor(shared): moved single-select to shared
