+++ src/shared/types/recruitment-status.ts
@@ -1,4 +1,4 @@
-import { Option } from './../../widgets/single-select/type';
+import { Option } from '@/shared/ui/single-select/type';
 
 export const recruitmentStatus: Option[] = [
 	{ value: 'recruitment-active', label: 'Набор открыт' },
