+++ src/shared/types/status-options.ts
@@ -1,4 +1,4 @@
-import { Option } from './../../widgets/single-select/type';
+import { Option } from '@/shared/ui/single-select/type';
 
 export const statusOptions: Option[] = [
 	{ value: 'active', label: 'Активный' },
