+++ src/shared/ui/single-select/type.ts
